Adding better server error handling (HTTP 500-504)
So as much as I'd like to believe that GitHub never errors, sometimes
it does.  Usually seen on larger repos, and a significant way into
the pagination.  This traps that there was an error, and sleeps for
a bit, and then tries again, up to 5 times.  5 is an arbitrary number
and seemed like a good idea

Signed-off-by: John 'Warthog9' Hawley <warthog9@eaglescrag.net>
Signed-off-by: John 'Warthog9' Hawley <jhawley@vmware.com>